refactored model image validation into concern
